introduce new rolloutcontroller interface
