+++ internal/controller/monovertexrollout/monovertexrollout_progressive.go
@@ -9,14 +9,12 @@ import (
 	apiv1 "github.com/numaproj/numaplane/pkg/apis/numaplane/v1alpha1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 
-	"sigs.k8s.io/controller-runtime/pkg/client"
-
 	"github.com/numaproj/numaplane/internal/common"
 	ctlrcommon "github.com/numaproj/numaplane/internal/controller/common"
 )
 
-// Implemented functions for the progressiveController interface:
-
+// CreateUpgradingChildDefinition creates a definition for an "upgrading" monovertex
+// This implements a function of the progressiveController interface
 func (r *MonoVertexRolloutReconciler) CreateUpgradingChildDefinition(ctx context.Context, rolloutObject ctlrcommon.RolloutObject, name string) (*unstructured.Unstructured, error) {
 	monoVertexRollout := rolloutObject.(*apiv1.MonoVertexRollout)
 	metadata, err := getBaseMonoVertexMetadata(monoVertexRollout)
@@ -35,56 +33,12 @@ func (r *MonoVertexRolloutReconciler) CreateUpgradingChildDefinition(ctx context
 	return monoVertex, nil
 }
 
-func (r *MonoVertexRolloutReconciler) getCurrentChildCount(rolloutObject ctlrcommon.RolloutObject) (int32, bool) {
-	monoVertexRollout := rolloutObject.(*apiv1.MonoVertexRollout)
-	if monoVertexRollout.Status.NameCount == nil {
-		return int32(0), false
-	} else {
-		return *monoVertexRollout.Status.NameCount, true
-	}
-}
-
-func (r *MonoVertexRolloutReconciler) updateCurrentChildCount(ctx context.Context, rolloutObject ctlrcommon.RolloutObject, nameCount int32) error {
-	monoVertexRollout := rolloutObject.(*apiv1.MonoVertexRollout)
-	monoVertexRollout.Status.NameCount = &nameCount
-	return r.updateMonoVertexRolloutStatus(ctx, monoVertexRollout)
-}
-
-// increment the child count for the Rollout and return the count to use
-func (r *MonoVertexRolloutReconciler) IncrementChildCount(ctx context.Context, rolloutObject ctlrcommon.RolloutObject) (int32, error) {
-	currentNameCount, found := r.getCurrentChildCount(rolloutObject)
-	if !found {
-		currentNameCount = int32(0)
-		err := r.updateCurrentChildCount(ctx, rolloutObject, int32(0))
-		if err != nil {
-			return int32(0), err
-		}
-	}
-
-	err := r.updateCurrentChildCount(ctx, rolloutObject, currentNameCount+1)
-	if err != nil {
-		return int32(0), err
-	}
-	return currentNameCount, nil
-}
-
-// Recycle deletes child; returns true if it was in fact deleted
-func (r *MonoVertexRolloutReconciler) Recycle(ctx context.Context,
-	monoVertexDef *unstructured.Unstructured,
-	c client.Client,
-) (bool, error) {
-	err := kubernetes.DeleteResource(ctx, c, monoVertexDef)
-	if err != nil {
-		return false, err
-	}
-	return true, nil
-}
-
 // AssessUpgradingChild makes an assessment of the upgrading child to determine if it was successful, failed, or still not known
 // Assessment:
 // Success: phase must be "Running" and all conditions must be True
 // Failure: phase is "Failed" or any condition is False
-// Unknowk: neither of the above if met
+// Unknown: neither of the above if met
+// This implements a function of the progressiveController interface
 // TODO: fix this assessment not to return an immediate result as soon as things are healthy or unhealthy
 func (r *MonoVertexRolloutReconciler) AssessUpgradingChild(ctx context.Context, existingUpgradingChildDef *unstructured.Unstructured) (apiv1.AssessmentResult, error) {
 	numaLogger := logger.FromContext(ctx)
