+++ internal/controller/progressive/progressive_test.go
@@ -99,7 +99,7 @@ func TestFindMostCurrentChildOfUpgradeState(t *testing.T) {
 				assert.NoError(t, err)
 			}
 
-			mostCurrentChild, err := FindMostCurrentChildOfUpgradeState(ctx, pipelineRollout, common.LabelValueUpgradePromoted, checkLive, client)
+			mostCurrentChild, err := ctlrcommon.FindMostCurrentChildOfUpgradeState(ctx, pipelineRollout, common.LabelValueUpgradePromoted, checkLive, client)
 			if tt.expectedError != nil {
 				assert.Error(t, err)
 				assert.Nil(t, mostCurrentChild)
